Random ids, flexible Thumbnails
Use large random integer keys for Photo and Event.
Store thumbnails in an outrigger table.
Re-organize S3 to use a photoboard bucket and not use the username or
file in the path.
Run the thumbnail generation in a goroutine.
Use a 940x705 thumbnail for the view photo page.
Add .CREDENTIALS to gitignore (AWS credentials)